StPicoDstMaker: Added new vertex selection criteria
Added new vertex selection mode "VpdOrDefault". In case there is no VPD vertex
available or no TPC vertex is close to the VPD one within a window the default
vertex (i.e. the first in the list) will be selected. This is particularly
important for triggers that do not require VPD coincidence, e.g. BHT2.