Use StPicoFmsFiller to fill FMS branch in picoDst
If the fmsDbMaker is available and the FMS hit branch in muDst is empty (i.e.
has zero entries) the hits are recovered from the corresponding data block in
StTriggerData. This is done to recover FMS hits in those muDst files in which
FMS hit branch was not filled as in Run 16 production.